refactor: TableEditableCell to be a separate component
