+++ packages/blade/src/components/Table/TableEditableCell.native.tsx
@@ -1,0 +1,9 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import type { TableEditableCellProps } from './types';
+import { Text } from '~components/Typography';
+
+const TableEditableCell = (props: TableEditableCellProps): React.ReactElement => {
+  return <Text>Table Component is not available for Native mobile apps.</Text>;
+};
+
+export { TableEditableCell };
