+++ packages/blade/src/components/Table/TableEditableCell.web.tsx
@@ -1,0 +1,143 @@
+import styled from 'styled-components';
+import { CellWrapper, StyledCell } from './TableBody';
+import { useTableContext } from './TableContext';
+import type { TableEditableCellProps, TableProps } from './types';
+import { tableEditableCellRowDensityToInputSizeMap } from './tokens';
+import { ComponentIds } from './componentIds';
+import { getFocusRingStyles } from '~utils/getFocusRingStyles';
+import { AlertCircleIcon, CheckIcon } from '~components/Icons';
+import type { MarginProps } from '~components/Box/BaseBox/types/spacingTypes';
+import { MetaConstants, metaAttribute } from '~utils/metaAttribute';
+import BaseBox from '~components/Box/BaseBox';
+import { Box } from '~components/Box';
+import { BaseInput } from '~components/Input/BaseInput';
+import { castWebType } from '~utils';
+import { assignWithoutSideEffects } from '~utils/assignWithoutSideEffects';
+
+export const StyledEditableCell = styled(StyledCell)<{
+  rowDensity: NonNullable<TableProps<unknown>['rowDensity']>;
+}>(({ theme, rowDensity }) => ({
+  '&&&': {
+    '&:focus-visible': { outline: '1px solid' },
+    '&:focus-within': {
+      ...(rowDensity !== 'comfortable' ? getFocusRingStyles({ theme, negativeOffset: true }) : {}),
+    },
+  },
+}));
+
+export const validationStateToInputTrailingIconMap = {
+  none: undefined,
+  success: CheckIcon,
+  error: AlertCircleIcon,
+};
+
+export const rowDensityToIsTableInputCellMapping = {
+  comfortable: false,
+  normal: true,
+  compact: true,
+};
+
+export const getEditableInputMargin = ({
+  rowDensity,
+}: {
+  rowDensity: NonNullable<TableProps<unknown>['rowDensity']>;
+}): MarginProps['margin'] => {
+  if (rowDensity === 'comfortable') {
+    return ['spacing.4', 'spacing.4'];
+  }
+
+  return 'spacing.2';
+};
+
+const _TableEditableCell = ({
+  validationState = 'none',
+  accessibilityLabel,
+  autoCapitalize,
+  autoCompleteSuggestionType,
+  autoFocus,
+  defaultValue,
+  isDisabled,
+  isRequired,
+  keyboardReturnKeyType,
+  leadingIcon,
+  maxCharacters,
+  name,
+  onBlur,
+  onChange,
+  onClick,
+  onFocus,
+  onSubmit,
+  placeholder,
+  prefix,
+  suffix,
+  value,
+  testID,
+  trailingButton,
+  errorText,
+  successText,
+}: TableEditableCellProps): React.ReactElement => {
+  const { rowDensity, showStripedRows, backgroundColor } = useTableContext();
+
+  return (
+    <StyledEditableCell
+      role="cell"
+      $backgroundColor={backgroundColor}
+      rowDensity={rowDensity}
+      {...metaAttribute({ name: MetaConstants.TableCell })}
+    >
+      <BaseBox className="cell-wrapper-base" display="flex" alignItems="center" height="100%">
+        <CellWrapper
+          className="cell-wrapper"
+          rowDensity={rowDensity}
+          showStripedRows={showStripedRows}
+          display="flex"
+          alignItems="center"
+          flex={1}
+          hasPadding={false}
+        >
+          <Box margin={getEditableInputMargin({ rowDensity })} width="100%">
+            <BaseInput
+              isTableInputCell={rowDensityToIsTableInputCellMapping[rowDensity]}
+              validationState={validationState}
+              id="table-editable-cell-input"
+              type="text"
+              size={tableEditableCellRowDensityToInputSizeMap[rowDensity]}
+              trailingIcon={validationStateToInputTrailingIconMap[validationState]}
+              accessibilityLabel={accessibilityLabel}
+              autoCapitalize={autoCapitalize}
+              autoCompleteSuggestionType={autoCompleteSuggestionType}
+              // eslint-disable-next-line jsx-a11y/no-autofocus
+              autoFocus={autoFocus}
+              defaultValue={defaultValue}
+              isDisabled={isDisabled}
+              isRequired={isRequired}
+              keyboardReturnKeyType={keyboardReturnKeyType}
+              leadingIcon={leadingIcon}
+              maxCharacters={maxCharacters}
+              name={name}
+              onBlur={onBlur}
+              onChange={onChange}
+              onClick={onClick}
+              onFocus={onFocus}
+              onSubmit={castWebType(onSubmit)}
+              placeholder={placeholder}
+              prefix={prefix}
+              suffix={suffix}
+              value={value}
+              testID={testID}
+              trailingButton={trailingButton}
+              errorText={errorText}
+              successText={successText}
+            />
+          </Box>
+        </CellWrapper>
+      </BaseBox>
+    </StyledEditableCell>
+  );
+};
+
+const TableEditableCell = assignWithoutSideEffects(_TableEditableCell, {
+  componentId: ComponentIds.TableEditableCell,
+});
+
+export { TableEditableCell };
