+++ packages/blade/src/components/Table/index.ts
@@ -5,3 +5,5 @@ export * from './TableFooter';
 export * from './TablePagination';
 export * from './TableToolbar';
 export * from './types';
+export { TableBody, TableCell, TableRow } from './TableBody';
+export { TableEditableCell } from './TableEditableCell';
